requirements.txt
asgiref==3.9.1
Django==4.2.7
django-cors-headers==4.7.0
djangorestframework==3.16.1
mysqlclient==2.2.7
sqlparse==0.5.3
tzdata==2025.2
